refactor(App): factor main App component into own module in components directory
